Update swap, add line counter
Start with an empty 'held piece' space; the first time we swap, put
the current piece into 'held', swap to the 'next' piece, and generate
a new next piece. Every time we swap after that, swap between the
'held' piece and the current piece.

Add a counter for lines cleared.