File refactoring!
Move 'Piece' class to its own file, along with closely associates,
CellValue and nullPiece. Kludge in setPieceField to set the 'field'
variable in the 'piece' module scope.

Move 'Field' class to its own file.

Move direction & command related definition to 'dir' to be used by
field and index.

tsconfig is along for the ride. It's currently configured to include
source maps and not really production suited, but this is still
all development code anyway.